Added documentation (in a comment in the file and here in the message, below) and removed extra commas (also explained in the documentation).
All activity photos are kept in the git repository now.
To add a new activity photo file, perform the following steps:
1. Put the file into: backend/web/static/images/activities
2. Add the filename and contributor to the JS structure below,
    make sure to include the prefix: path +

When adding a URL, look at one of the http / https examples below.
For a URL, the path + prefix must NOT be added.

In PHP it is recommended to add a comma on the last array element.
In JS files (like this one) an extra comma results in inconsistent
array lengths in different browsers and is therefore NOT recommended.